Make resize-images script look into directories recursively

+++ src/lib/utils/resize-images.js
@@ -6,38 +6,49 @@ const directoryPath = join(process.cwd(), 'static');
 const maxWidth = 1080;
 const maxHeight = 1080;
 
-readdir(directoryPath, (err, files) => {
-  if (err) {
-    return console.log('Unable to scan directory: ' + err);
-  }
+/**
+ * @param {string} directoryPath
+ */
+function processDirectory(directoryPath) {
+  readdir(directoryPath, { withFileTypes: true }, (err, files) => {
+    if (err) {
+      return console.log('Unable to scan directory: ' + err);
+    }
 
-  files.forEach((file) => {
-    const filePath = join(directoryPath, file);
-    const tempFilePath = `${filePath}_temp`;
+    files.forEach((file) => {
+      const filePath = join(directoryPath, file.name);
+      if (file.isDirectory()) {
+        processDirectory(filePath);
+      } else {
+        const tempFilePath = `${filePath}_temp`;
 
-    sharp(filePath)
-      .resize(maxWidth, maxHeight, {
-        fit: 'inside',
-        withoutEnlargement: true,
-      })
-      .toFile(tempFilePath, (err) => {
-        if (err) {
-          console.log('Error resizing image: ', err);
-        } else {
-          unlink(filePath, (err) => {
+        sharp(filePath)
+          .resize(maxWidth, maxHeight, {
+            fit: 'inside',
+            withoutEnlargement: true,
+          })
+          .toFile(tempFilePath, (err) => {
             if (err) {
-              console.log('Error deleting original file: ', err);
+              console.log('Error resizing image: ', err);
             } else {
-              rename(tempFilePath, filePath, (err) => {
+              unlink(filePath, (err) => {
                 if (err) {
-                  console.log('Error renaming temporary file: ', err);
+                  console.log('Error deleting original file: ', err);
                 } else {
-                  console.log(`Resized image saved as ${filePath}`);
+                  rename(tempFilePath, filePath, (err) => {
+                    if (err) {
+                      console.log('Error renaming temporary file: ', err);
+                    } else {
+                      console.log(`Resized image saved as ${filePath}`);
+                    }
+                  });
                 }
               });
             }
           });
-        }
-      });
+      }
+    });
   });
-});
+}
+
+processDirectory(directoryPath);
